requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
Django==5.1
django-environ==0.11.2
django-phonenumber-field==8.0.0
phonenumbers==8.13.42
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.1
